UPSTREAM: drm/i915/guc: Insert submit fences between requests in parent-child relationship
The GuC must receive requests in the order submitted for contexts in a
parent-child relationship to function correctly. To ensure this, insert
a submit fence between the current request and last request submitted
for requests / contexts in a parent child relationship. This is
conceptually similar to a single timeline.

Signed-off-by: Matthew Brost <matthew.brost@intel.com>
Cc: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: John Harrison <John.C.Harrison@Intel.com>
Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20211014172005.27155-14-matthew.brost@intel.com
(cherry picked from commit bc955204919ea8152b7443e7d48a48cc18dea448)

BUG=b:239109915
TEST=Boot and test graphics on Brya or Skolas

Signed-off-by: Shinni Tsai <shinni.tsai@intel.com>
Change-Id: I15a4ccba97cd49822b0deba7f99ddc1d4f7da5e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4286673
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>