UPSTREAM: drm/i915: Make request conflict tracking understand parallel submits
If an object in the excl or shared slot is a composite fence from a
parallel submit and the current request in the conflict tracking is from
the same parallel context there is no need to enforce ordering as the
ordering is already implicit. Make the request conflict tracking
understand this by comparing a parallel submit's parent context and
skipping conflict insertion if the values match.

v2:
 (John Harrison)
  - Reword commit message

Signed-off-by: Matthew Brost <matthew.brost@intel.com>
Reviewed-by: John Harrison <John.C.Harrison@Intel.com>
Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20211014172005.27155-23-matthew.brost@intel.com
(cherry picked from commit afc76f307e60c865c436e3828a7756e0c358fe0d)

BUG=b:239109915
TEST=Boot and test graphics on Brya or Skolas

Signed-off-by: Shinni Tsai <shinni.tsai@intel.com>
Change-Id: I6fb8ce17255c0d9ded1e95637e3eec6af3aaf0f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4286682
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>