CHROMIUM: rk3399: Sync with vblank in kernel for ddrfreq
This is based off the chromeos-4.4 branch ToT, rather than porting a
bunch of individual commits, rewrites, and fixups. It is effectively a
squash of commits like the following, and probably a few other minor
ones I've not listed specifically:

f2e0e3cb3993 CHROMIUM: rk3399: Sync with vblank in the kernel for ddrfreq
1d8ab1ed92c8 CHROMIUM: drm/rockchip: Remove locks for dmcfreq handling
e2cc20adc4fd CHROMIUM: rk3399: Disable ddr clk timeout on suspend
599e97d6eacf CHROMIUM: rk3399: dmc: fix enable/disable issue
d5351b5a46c2 CHROMIUM: devfreq: rk3399: Fix dmcfreq unblock unbalanced

<original description>
This changes the kernel to sync with vblank for the display in the
kernel. This was done in ARM Trusted Firmware before, but that locks
up one CPU for up to one display frame (1/60 second). That's bad for
performance and power, so this moves waiting to the kernel where the
waiting thread can properly wait on a completion.
</original description>

Non-trivial porting notes:

* reintroduce rockchip_drm_atomic_commit()
* replace for_each_old_crtc_in_state() with
  for_each_oldnew_crtc_in_state() in block/unblock checks
* careful arrangement of rockchip_dmcfreq_unregister_clk_sync_nb()
  within vop_crtc_atomic_disable(), now that PSR support is embedded in
  the atomic enable()/disable() functions
* handle priv->devfreq==NULL in rockchip_drm_vop.c
* switch to SPDX headers, for new files
* move drivers/devfreq/rk3399_dmc_priv.h ->
  include/soc/rockchip/rk3399_ddrclk.h, so both declaration and usage
  can reference it properly

Upstreaming notes: this is CHROMIUM, because it's not a very
clean integration of DRM and clock drivers. Ideally, we could offload
this asynchronous clock+vblank management to ARM Trusted Firmware (where
it should then offload the vblank wait to a low-power M0 core, rather
than occupying a full CPU), and then this patch becomes much simpler --
just an asynchronous ATF call.

BUG=b:202883955
TEST=watch /sys/class/devfreq/dmc/cur_freq
TEST=switch governors (e.g., user space)
TEST=watch for cursor jank

Change-Id: I3f6407d77dc6a67455ce774c1560afc630a7b04c
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3368886