FROMGIT: drm/msm/dp: Modify prototype of encoder based API
Functions in the DisplayPort code that relates to individual instances
(encoders) are passed both the struct msm_dp and the struct drm_encoder.
But in a situation where multiple DP instances would exist this means
that the caller need to resolve which struct msm_dp relates to the
struct drm_encoder at hand.

Store a reference to the struct msm_dp associated with each
dpu_encoder_virt to allow the particular instance to be associate with
the encoder in the following patch.

Reviewed-by: Abhinav Kumar <abhinavk@codeaurora.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20211016221843.2167329-3-bjorn.andersson@linaro.org
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 167dac97eb46c7b8a15b2195080e191bb0c9ce35
 https://gitlab.freedesktop.org/drm/msm.git msm-next)

BUG=b:201590452, b:177187240
TEST=emerge-herobrine chromeos-kernel-5_10

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I88a68cf81741ab0d4eb82cb2f1dab927634b9013
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3253992
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>