UPSTREAM: platform/x86: amd-pmc: adjust arguments for `amd_pmc_send_cmd`
Currently the "argument" for the "message" is listed as a boolean
value.  This works well for the commands used currently, but an
additional upcoming command will pass more data in the message.

Expand it to be a full 32 bit value.

Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Link: https://lore.kernel.org/r/20211020162946.10537-1-mario.limonciello@amd.com
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
(cherry picked from commit 4c9dbf8622797eee44e9a4f7327d3d04aa4d1d5f)

BUG=b:194146855
TEST=Build and deploy kernel

Signed-off-by: Deepak Sharma <deepak.sharma@amd.com>
Change-Id: I0bf88681f60a78a181636f806c7cc8e810f5e480
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3292326
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rajib Mahapatra <rajib.mahapatra@amd.corp-partner.google.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Rajib Mahapatra <rajib.mahapatra@amd.corp-partner.google.com>