UPSTREAM: platform/x86: amd-pmc: Add special handling for timer based S0i3 wakeup
RTC based wakeup from s0i3 doesn't work properly on some Green Sardine
platforms. Because of this, a newer SMU for Green Sardine has the ability
to pass wakeup time as argument of the upper 16 bits of OS_HINT message.

With older firmware setting the timer value in OS_HINT will cause firmware
to reject the hint, so only run this path on:
1) Green Sardine
2) Minimum SMU FW
3) RTC alarm armed during s0i3 entry

Using this method has some limitations that the s0i3 wakeup will need to
be between 4 seconds and 18 hours, so check those boundary conditions as
well and abort the suspend if RTC is armed for too short or too long of a
duration.

Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Link: https://lore.kernel.org/r/20211020162946.10537-2-mario.limonciello@amd.com
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
(cherry picked from commit 59348401ebed9f0e8ffe2d5b9cf1de30ecb24dde)

BUG=b:194146855
TEST=Build and deploy kernel

Signed-off-by: Deepak Sharma <deepak.sharma@amd.com>
Change-Id: I220168d81df95ff47fc81a120fcb46fc3c8851ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3292327
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rajib Mahapatra <rajib.mahapatra@amd.corp-partner.google.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Rajib Mahapatra <rajib.mahapatra@amd.corp-partner.google.com>