CHROMIUM: drm/i915/gt: Spread virtual engines over idle engines
Everytime we come to the end of a virtual engine's context, re-randomise
it's siblings[]. As we schedule the siblings' tasklets in the order they
are in the array, earlier entries are executed first (when idle) and so
will be preferred when scheduling the next virtual request. Currently,
we only update the array when switching onto a new idle engine, so we
prefer to stick on the last execute engine, keeping the work compact.
However, it can be beneficial to spread the work out across idle
engines, so choose another sibling as our preferred target at the end of
the context's execution.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
(am from https://patchwork.freedesktop.org/series/97043/)

BUG=b:206601681, b:214267861
TEST=video play back and run intel_gpu_top -s 10

Signed-off-by: George D Sworo <george.d.sworo@intel.com>
Change-Id: Idb18cdd575cf196aa1f845a08af77a786999c269
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3319241
Reviewed-by: Drew Davenport <ddavenport@chromium.org>