UPSTREAM: scsi: ufs: core: Fix another task management completion race
hba->outstanding_tasks, which is read under host_lock spinlock, tells the
interrupt handler what task management tags are in use by the driver.  The
doorbell register bits indicate which tags are in use by the hardware.  A
doorbell bit that is 0 is because the bit has yet to be set by the driver,
or because the task is complete. It is only possible to disambiguate the 2
cases, if reading/writing the doorbell register is synchronized with
reading/writing hba->outstanding_tasks.

For that reason, reading REG_UTP_TASK_REQ_DOOR_BELL must be done under
spinlock.

Link: https://lore.kernel.org/r/20211108064815.569494-3-adrian.hunter@intel.com
Fixes: f5ef336fd2e4 ("scsi: ufs: core: Fix task management completion")
Reviewed-by: Bart Van Assche <bvanassche@acm.org>
Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 5cb37a26355d79ab290220677b1b57d28e99a895)

BUG=b:245912634
TEST=build and check driver bind

Signed-off-by: Mohandass, Divagar <divagar.mohandass@intel.com>
Change-Id: Iee0b37a1120b33d3017aa439d39086bdc0ddaed2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3904714
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>