FROMLIST: dt-bindings: mediatek,dpi: Add DP_INTF compatible
DP_INTF is similar to DPI but does not have the exact same feature set
or register layouts.

DP_INTF is the sink of the display pipeline that is connected to the
DisplayPort controller and encoder unit. It takes the same clocks as
DPI.
(am from https://patchwork.freedesktop.org/patch/462696/)

BUG=b:205808556
TEST=Boot to home.

Signed-off-by: Jitao Shi <jitao.shi@mediatek.com>
Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Guillaume Ranquet <granquet@baylibre.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Change-Id: Iaa4c8174a82bd8c91255e0ed0a33fe2badd519ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3292848
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Tested-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>