UPSTREAM: mt76: mt7915: fix hwmon temp sensor mem use-after-free
Without this change, garbage is seen in the hwmon name and sensors output
for mt7915 is garbled. It appears that the hwmon logic does not make a
copy of the incoming string, but instead just copies a char* and expects
it to never go away.

Fixes: 33fe9c639c13 ("mt76: mt7915: add thermal sensor device support")
Signed-off-by: Ben Greear <greearb@candelatech.com>
Signed-off-by: Ryder Lee <ryder.lee@mediatek.com>
Signed-off-by: Felix Fietkau <nbd@nbd.name>
(cherry picked from commit 0ae3ff5684514d72357240f1033a7494c51f93ed)

BUG=b:178754244, b:188614537
TEST=Merge patch from upstream

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: I0b7f7ab24af4735f5aee26d56fc135200e462e93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3281385
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Joshua Emele <jemele@chromium.org>
Commit-Queue: Joshua Emele <jemele@chromium.org>
Tested-by: Joshua Emele <jemele@chromium.org>