UPSTREAM: firmware: cs_dsp: Add version checks on coefficient loading
The firmware coefficient files contain version information that is
currently ignored by the cs_dsp code. This information specifies which
version of the firmware the coefficient were generated for. Add a check
into the code which prints a warning in the case the coefficient and
firmware differ in version, in many cases this will be ok but it is not
always, so best to let the user know there is a potential issue.

Co-authored-by: Simon Trimmer <simont@opensource.cirrus.com>
Signed-off-by: Simon Trimmer <simont@opensource.cirrus.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20211117132300.1290-3-ckeepax@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 2925748eadc33cba3bded7b69475a1b002b124ac)

BUG=b:186972464
TEST=ODM tests audio function on Vell

Signed-off-by: Brent Lu <brent.lu@intel.com>
Change-Id: I0ac85583db7d6eadc2715995ae659d2d544a04a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3471992
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Shon Wang <shon.wang@quanta.corp-partner.google.com>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>