CHROMIUM: drm/input_helper: Add new input-handling helper
A variety of applications have found it useful to listen to
user-initiated input events to make decisions within a DRM driver, given
that input events are often the first sign that we're going to start
doing latency-sensitive activities:

 * Panel self-refresh: software-directed self-refresh (e.g., with
   Rockchip eDP) is especially latency sensitive. In some cases, it can
   take 10s of milliseconds for a panel to exit self-refresh, which can
   be noticeable. Rockchip RK3399 Chrome OS systems have always shipped
   with an input_handler boost, that preemptively exits self-refresh
   whenever there is input activity.

 * GPU drivers: on GPU-accelerated desktop systems, we may need to
   render new frames immediately after user activity. Powering up the
   GPU can take enough time that it is worthwhile to start this process
   as soon as there is input activity. Many Chrome OS systems also ship
   with an input_handler boost that powers up the GPU.

This patch provides a small helper library that abstracts some of the
input-subsystem details around picking which devices to listen to, and
some other boilerplate. This will be used in the next patch to implement
the first bullet: preemptive exit for panel self-refresh.

Bits of this are adapted from code the Android and/or Chrome OS kernels
have been carrying for a while.

Signed-off-by: Brian Norris <briannorris@chromium.org>
(am from https://patchwork.freedesktop.org/patch/463497/)
(also found at https://lore.kernel.org/r/20211117144807.v2.1.I09b516eff75ead160a6582dd557e7e7e900c9e8e@changeid)

CHROMIUM notes:
This was submitted verbatim upstream to mixed reviews. In short, this is
most likely not acceptable upstream, primarily because several folks
would prefer the policy aspects (around which input device(s) to look
at, and whether a given DRM device warrants an input boost or not) be
determined by user space. That would be a moderately complex rewrite,
and this is simple, so I skip that.

Perhaps the simplest way forward upstream would be to expose more
mechanisms as ioctls (e.g., the PSR-exit that uses this in the next CL),
and listen from a daemon (e.g., powerd). But I'll call that Future
Work(TM).

BUG=b:172220334
TEST=PSR on kevin (next patch)

Change-Id: I09b516eff75ead160a6582dd557e7e7e900c9e8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3260953
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>