soc/google/cpif: Refactor the busy flag of tx flowctrl
"under_flow_ctrl" checks the busy flag of queues. The flag is set when
a queue is full but the name has "flow_ctrl" that CP requests.

Rename "under_flow_ctrl" to "check_busy" and move the checking busy
flag into "check_busy". It can simplify the checking routine in
link_device.c and wrap the flag in netif_lock like the other usages.

Change-Id: I63a1064f3e3267fdd565b2e156691259278272e0
Signed-off-by: Dongseok Yi <dseok.yi@samsung.com>