HACK: CHROMIUM: media: hantro: Compute plane address with data_offset
(reformatted but otherwise cargo-culted commit message)

With V4L2 MEMORY DMABUF for Multiplane API, it is necessary to set a
file descriptor on each plane. The way to pass an offset with a
DMA-buf is not defined in V4L2 specification. Some drivers need to
abuse the data_offset field for the purpose.

This patch fills the data_offset field of vb2_plane by the same field
of v4l2_plane, so drivers may utilize it when needed.

BUG=chromium:901264
BUG=b:172216855
BUG=b:206738124
TEST=Run jpeg_encode_accelerator_unittests on kevin-kernelnext

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: Ic2c83bbde7ec0386398a1945fea2d93b6be97c68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3291289
Reviewed-by: Brian Norris <briannorris@chromium.org>