BACKPORT: kmap_local: don't assume kmap PTEs are linear arrays in memory
The kmap_local conversion broke the ARM architecture, because the new
code assumes that all PTEs used for creating kmaps form a linear array
in memory, and uses array indexing to look up the kmap PTE belonging to
a certain kmap index.

On ARM, this cannot work, not only because the PTE pages may be
non-adjacent in memory, but also because ARM/!LPAE interleaves hardware
entries and extended entries (carrying software-only bits) in a way that
is not compatible with array indexing.

Fortunately, this only seems to affect configurations with more than 8
CPUs, due to the way the per-CPU kmap slots are organized in memory.

Work around this by permitting an architecture to set a Kconfig symbol
that signifies that the kmap PTEs do not form a lineary array in memory,
and so the only way to locate the appropriate one is to walk the page
tables.

Link: https://lore.kernel.org/linux-arm-kernel/20211026131249.3731275-1-ardb@kernel.org/
Link: https://lkml.kernel.org/r/20211116094737.7391-1-ardb@kernel.org
Fixes: 2a15ba82fa6c ("ARM: highmem: Switch to generic kmap atomic")
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reported-by: Quanyang Wang <quanyang.wang@windriver.com>
Reviewed-by: Linus Walleij <linus.walleij@linaro.org>
Acked-by: Russell King (Oracle) <rmk+kernel@armlinux.org.uk>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

(cherry picked from commit 825c43f50e3aa811a291ffcb40e02fbf6d91ba86)

Conflicts:
   mm/Kconfig
[ Different context.  See b/222221770#comment2. ]

BUG=b:222221770
BUG=b:195001087
BUG=b:191366271
TEST=for arch in amd64-generic arm64-generic; do \
        USE=buildtest emerge-"${arch}" chromeos-kernel-5_10; \
     done

Signed-off-by: Tzung-Bi Shih <tzungbi@chromium.org>
Change-Id: I39ffb5f407939d7991a10738bde132582bb3f5f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3500322
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>