CHROMIUM: iwl7000: mvm: Correctly set fragmented EBS
Currently, fragmented EBS was set for a channel only if the 'hb_type'
was set to fragmented or balanced scan. However, 'hb_type' is set only
in case of CDB, and thus fragmented EBS is never set for a channel for
non-CDB devices. Fix it.

Change-Id: I654eafa62fd647030ae6d4f07f32c96c3171decb
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: 0f8c865bd8076776327fe351cf390bf5f467cc6c