CHROMIUM: coiommu: introduce a global coiommu data structure
global_coiommu is a data structure which will be initialized only
when there is a coiommu device enumerated during kernel booting. If
coiommu device is detected, global_coiommu will be initialized with the
topology information from the device. This is done before the coiommu
driver probing. With such topology info, coiommu can further setup the
dma_map_ops for the devices in its control.

BUG=b:188481989
TEST=build kernel with coiommu config enabled

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.corp-partner.google.com>
Signed-off-by: Yu Zhang <yu.c.zhang@linux.intel.com>
Change-Id: I3bf110b4e820f64338630dcf95c39603fef7569d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3292615
Reviewed-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>