CHROMIUM: coiommu: initialize the DMA tracking table
DMA tracking table(DTT) will be used by coiommu to track the map and pin
status for each pfn. It is a page-table like structure with a root
page pointer and level. This will be initialized during coiommu device
driver probing as the root pointer and level will also be needed by
the coiommu backend. The DTT will be used in the next patches.

BUG=b:188481989
TEST=build kernel with coiommu config enabled

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.corp-partner.google.com>
Signed-off-by: Yu Zhang <yu.c.zhang@linux.intel.com>
Change-Id: I758ca4c5af0cfbeb987b2d606d8ca9a1ece65d80
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3292616
Reviewed-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>