CHROMIUM: coiommu: add initial coiommu functionalities
CoIOMMU provides a DMA map ops to the device driver. This ops uses use
identical mapping between IOVA and GPA. This can help to reduce the
pin/unpin request count sending to the backend through the pin ops. With
this, once a guest page is pinned by coiommu backend then no need to
re-pin this page again due to allocating a different IOVA for the same GPA.

To identify if a page is pinned or not, coiommu marks pfn with different
flags which will be introduced in the next patches.

BUG=b:188481989
TEST=build kernel with coiommu config enabled

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.corp-partner.google.com>
Signed-off-by: Yu Zhang <yu.c.zhang@linux.intel.com>
Change-Id: Ib729fd3a4a3a39f9008c9db0d0215c3e81e6c05c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3292613
Reviewed-by: David Stevens <stevensd@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>