CHROMIUM: drm/panfrost: nerf driver by changing compat string
For various reasons, some of which were discussed in [1], we
want both the vendor and panfrost Mali drivers be enabled in
the CrOS kernel trees.

Currently the vendor probe() succeeds and panfrost fails due
to "CHROMIUM: arm64: dts: mt8183: Add gpu node and opp lists"
but relying on these DT differences to determine the driver
is a bit risky, so we deliberately nerf panfrost.

Those who wish to experiment with panfrost on specific boards
will have to explicitly change the compat. When / if ChromeOS
is ready for panfrost by default this should be reverted and
the vendor-specific bits deleted.

[1] https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3311282

BUG=b:205549284
TEST=Local builds & boot on jacuzzi (MT8183); CQ.

Change-Id: I2b1af1dd15467376eb3d8af77f5dc95ad172affa
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3329884
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>