FROMLIST: x86: intel_epb: Allow model specific normal EPB value
The current EPB "normal" is defined as 6 and set whenever power-up EPB
value is 0. This setting resulted in the desired out of box power and
performance for several CPU generations. But this value is not suitable
for AlderLake mobile CPUs, as this resulted in higher uncore power.
Since EPB is model specific, this is not unreasonable to have different
behavior.

Allow a capability where "normal" EPB can be redefined. For AlderLake
mobile CPUs this desired normal value is 7.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
(am from
https://lore.kernel.org/lkml/20211211163327.3093169-1-srinivas.pandruvada@linux.intel.com/)

BUG=b:209393800, b:209626955
TEST= boot and read msr: rdmsr -a 0x1b0

Signed-off-by: George D Sworo <george.d.sworo@intel.com>
Change-Id: Ie91c2273d0f4d74c3c4bf776dfc505e2afcf04c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3318530
Reviewed-by: Rajat Jain <rajatja@google.com>