CHROMIUM: gpu: mali: Refactor platform code
Referenced from CL:2847956.

MTK boards on the same kernel version share the same splitconfig, which
make it hard to add link to their own Mali platform configs.

This CL adapts how the core Mali driver access the platform specific
config using compatible strings, so we can better address the issue
above.

This is meant to be the base change that only adapts the native
"devicetree" platform, so future uprev should be able to pick up the
change easier. The target platform code will be added later based on
this CL.

BUG=b:207573982,b:186163221
TEST=CONFIG_MALI_VALHALL_PLATFORM_NAME="devicetree" in MTK flavoured
     config
     emerge-cherry chromeos-kernel-5_10

Change-Id: I7e07b55e1737f5d3980dd733c29687ecf4d49dff
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3334329
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>