CHROMIUM: arm64: dts: mt8195: cherry: Fix up display panel regulator
The regulator for the display panel was added without an upstream
supply, incorrect name, and incorrect description.

The regulator is a simple power switch. The voltage should be derived
from its upstream supply.

Now that the system power rails have been added, fix up the power switch
for the display panel. Correct the name so that it matches the rail name
on the schematics, drop the voltage constraints, and add its supply.

BUG=b:188609319
BUG=b:189297520
TEST=cat /sys/kernel/debug/regulator/regulator_summary | grep pp[0-9]
     "pp3300_disp_x" (this one) should be consumer of "pp3300_z2".

Fixes: 2eabcb52f84e ("CHROMIUM: arm64: dts: mt8195: add eDP nodes")
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: I5ad2868e2f7999c9bac0d5439583d16be8957893
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3344472
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>