CHROMIUM: arm64: dts: mt8195: cherry: add regulator for touchscreen
The touchscreen is missing its power supply in the device tree. It is
powered from the switched 3.3V system power rail.

Add it to the device tree to correctly describe the hardware. On Linux
this gets rid of the regulator supply not found message.

>> i2c_hid_of 4-0010: supply vdd not found, using dummy regulator

The kernel also complains about vddl not found. But according to the
bindings, that is only required by one model, and not the model this
device uses. To fix that the driver would need to be fixed.

BUG=b:185860653
BUG=b:188609319
BUG=b:209733810
TEST=dmesg | grep supply

Fixes: 3dbe0c794d74 ("CHROMIUM: arm64: dts: mt8195: support touchscreen")
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: I59ad6f00598674f2fd0262c1813b8cabc889e367
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3344475
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>