CHROMIUM: arm64: dts: mt8195: cherry: add regulator for audio codec
The RT5682 audio codec is missing its power supplies in the device tree.
These are all powered from system power rails without additional
switches. Curiously, both VBAT and MICVDD are provided, even though it
seems only one might actually be needed. Internally, VBAT feeds LDO2
to provide MICVDD.

For the RT5682S, VBAT is not connected internally, and is not needed.

Add them to the device tree to correctly describe the hardware. On Linux
this gets rid of regulator supplies not found messages.

>> rt5682 2-001a: supply AVDD not found, using dummy regulator
>> rt5682 2-001a: supply MICVDD not found, using dummy regulator
>> rt5682 2-001a: supply VBAT not found, using dummy regulator

BUG=b:188609319
BUG=b:188612929
BUG=b:199485457
BUG=b:209733810
TEST=dmesg | grep supply

Fixes: 5013243ee085 ("CHROMIUM: arm64: dts: mt8195: enable audio related nodes")
Fixes: 11c5eb80a10c ("CHROMIUM: arm64: dts: mt8195: add rt5682s node to i2c2 node")
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: I74093aa8393b0efa54f8cf0eccfb4868108ea516
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3344476
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>