FROMLIST: platform/chrome: Add driver for ChromeOS privacy-screen
This adds the ACPI driver for the ChromeOS privacy screen that is
present on some chromeos devices.

Note that ideally, we'd want this privacy screen driver to be probed
BEFORE the drm probe in order to avoid a drm probe deferral:
https://hansdegoede.livejournal.com/25948.html

In practise, I found that ACPI drivers are bound to their devices AFTER
the drm probe on chromebooks. So on chromebooks with privacy-screen,
this patch along with the other one in this series results in a probe
deferral of about 250ms for i915 driver. However, it did not result in
any user noticeable delay of splash screen in my personal experience.

In future if this probe deferral turns out to be an issue, we can
consider turning this ACPI driver into something that is probed
earlier than the drm drivers.

Signed-off-by: Rajat Jain <rajatja@google.com>
(am from https://patchwork.kernel.org/patch/12690727/)
(also found at https://lore.kernel.org/r/20211222001127.3337471-2-rajatja@google.com)

BUG=b:196060866
TEST=Test privacy-screen on Redrix

Change-Id: I85b94e9d4140f6745e6f0b071571347dd4f69bd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3324351
Tested-by: Rajat Jain <rajatja@google.com>
Auto-Submit: Rajat Jain <rajatja@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Rajat Jain <rajatja@google.com>