Revert "CHROMIUM: drm/i915: Wait for vblank after disabling cursor"
This reverts commit 1cca046cc0b436dccb12da9b57c754b7ce2b3a99.

Reason for revert:
This commit has been backported since 4.4 gen9 kernels.
This change is not in upstream and i915 developers don't see a need for
it anymore for gen12+

Reverting this also fixes several issues with PSR flickering and 8k
cursor lag.
The issue mentioned in the original commit was not reproducible on
brya.

BUG=b:209572830,b:209531196,b:210542797
TEST=No flickering when cursor moves over video playback

Signed-off-by: Juston Li <juston.li@intel.com>
Change-Id: Ifbcde0fcec5ecdec84b2ed2fa77b6294468e6438
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3354898
Reviewed-by: Vidya Srinivas <vidya.srinivas@intel.com>
Tested-by: Vidya Srinivas <vidya.srinivas@intel.com>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>