FROMGIT: ASoC: rt5682: Register wclk with its parent_hws instead of parent_data
The mclk might not be registered as a fixed clk name "mclk" on some
platforms.
In those platforms, if the mclk needed to be controlled by codec driver
and acquired by a fixed name, it would be a problem.

This patch to fix the issue that wclk becomes an orphan due to the fixed
mclk's name.

Signed-off-by: Derek Fang <derek.fang@realtek.com>
Link: https://lore.kernel.org/r/20211227055446.27563-1-derek.fang@realtek.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit cc5c9788106fb1b9e03c8c57d8d7166073a54416
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git for-next)

BUG=b:211531991
TEST=tested by LCFC/Realtek on Grunt

Signed-off-by: Derek Fang <derek.fang@realtek.corp-partner.google.com>
Change-Id: Ie62685e8fe9cab677906f0ab47f15e9f1de1637a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3361496
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>