CHROMIUM: drm: i915: Quirk: Swap opregion version fields
In 2017 Coreboot patch "UPSTREAM: drv/intel/gma/opregion:
fix opregion version for Windows" modifies the opregion
field format in order to fix the Windows bug.

The comment to this patch says: As the Linux i915 driver
only checks that the version is >= 2, there is no change in
functionality there.

This was true till kernel v5.2 where introduction of patch:
"drm/i915/opregion: fix version check" assumed we will use
the fixed Coreboot (first mentioned patch).

In reality FW is updated very rarely but we update the
kernel in the kernel uprev process. As result when crossing
kernel version 5.2 on platforms with old Coreboot creates a
problem with HDMI as the DRM code fails to find VBIOS
tables.

This patch adds quirk that checks if the version field in
opregion header is in reversed format and would result in
incorrect version reading. It swaps the bytes as they would
be produced by fixed Coreboot code.

BUG=b:207769981
TEST=Build Coral image for kernelnext (5.10) and verify HDMI
     is working. You may also look into dmesg
     dmesg | grep "swapping fields in opregion"

Change-Id: If2d53806b15bc52ec3b9badf75965fc171363744
Signed-off-by: Radoslaw Biernacki <rad@semihalf.com>
Signed-off-by: Lukasz Majczak <lma@semihalf.com>
Signed-off-by: marek Maslanka <mm@semihalf.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3358200
Tested-by: Radoslaw Biernacki <biernacki@google.com>
Reviewed-by: Marcin Wojtas <mwojtas@google.com>
Commit-Queue: Marcin Wojtas <mwojtas@google.com>