UPSTREAM: ASoC: mediatek: mt8195: correct pcmif BE dai control flow
Originally, the conditions for preventing reentry are not correct.
dai->component->active is not the state specifically for pcmif dai, so it
is not a correct condition to indicate the status of pcmif dai.
On the other hand, snd_soc_dai_stream_actvie() in prepare ops for both
playback and capture possibly return true at the first entry when these
two streams are opened at the same time.

In the patch, I refer to the implementation in mt8192-dai-pcm.c.
Clock and enabling bit for PCMIF are managed by DAPM, and the condition
for prepare ops is replaced by the status of dai widget.

Fixes: 1f95c019115c ("ASoC: mediatek: mt8195: support pcm in platform driver")
Signed-off-by: Trevor Wu <trevor.wu@mediatek.com>
Link: https://lore.kernel.org/r/20211230084731.31372-2-trevor.wu@mediatek.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 2355028c0c54c03afb66c589347f1dc9f6fe2e38)

BUG=b:204391159
TEST=emerge-cherry sys-kernel/chromeos-kernel-5_10

Change-Id: I41438f2d59dccd1cc7d967cb2dc7ea73fccfe4d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3506028
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: Trevor Wu <trevor.wu@mediatek.corp-partner.google.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>