UPSTREAM: firmware: cs_dsp: Clear core reset for cache
If the Halo registers are kept in the register cache the
HALO_CORE_RESET bit will be retained as 1 after reset is triggered in
cs_dsp_halo_start_core. This will cause subsequent writes to reset
the core which is not desired. Apart from this bit the rest of the
register bits are cacheable, so for safety sake clear the bit to
ensure the cache is consistent.

Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20220105113026.18955-6-ckeepax@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 7aa1cc1091e0a424e9e7711ca381ebe98b6865bc)

BUG=b:186972464
TEST=ODM tests audio function on Vell

Signed-off-by: Brent Lu <brent.lu@intel.com>
Change-Id: I1428c097f29b20e4c3bfe8a11679e0403d914a9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3472001
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Shon Wang <shon.wang@quanta.corp-partner.google.com>
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Commit-Queue: Terry Cheong <htcheong@chromium.org>