UPSTREAM: ASoC: wm_adsp: Add support for "toggle" preloaders
In the case a device can support retaining the firmware memory across
low power states it is useful for the preloader widget to only power up
whilst actually loading/unloading the core, as opposed to the normal
operation where the widget is powered for the entire time a firmware is
preloaded onto the core. Add support for this mode and a flag to enable
it.

Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20220105113026.18955-7-ckeepax@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit ba235634b138cd9d012dbe983e7920481211e132)

BUG=b:186972464
TEST=ODM tests audio function on Vell

Signed-off-by: Brent Lu <brent.lu@intel.com>
Change-Id: I39063956138bdaeb7799cd24a33da90c4dfa4bd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3472002
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Shon Wang <shon.wang@quanta.corp-partner.google.com>
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Commit-Queue: Terry Cheong <htcheong@chromium.org>