BACKPORT: FROMGIT: Bluetooth: mt7921s: Support wake on bluetooth
Enable wake on bluetooth on mt7921s that can be supported since the
firmware with version 20211129211059 was added, and the patch would
not cause any harm even when the old firmware is applied.

The patch was tested by setting up an HID or HOGP profile to connect a
Bluetooth keyboard and mouse, then putting the system to suspend, then
trying to wake up the system by moving the Bluetooth keyboard or mouse,
and then checking if the system can wake up and be brought back to
the normal state.

Co-developed-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Mark Chen <mark-yw.chen@mediatek.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit ce64b3e949195768361b0e4111e0837e5e46930d
 https://git.kernel.org/pub/scm/linux/kernel/git/bluetooth/bluetooth-next.git master)

Backport Notes:
	backporting because "Bluetooth: Rename driver .prevent_wake to .wakeup"
	is not included

BUG=b:178754244
TEST=build on cherry

Signed-off-by: Sean Wang <objelf@gmail.com>
Change-Id: I9a921be8030a154a09d8aca2885514adce369578
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3397938
Reviewed-by: Joshua Emele <jemele@chromium.org>
Tested-by: Joshua Emele <jemele@chromium.org>
Commit-Queue: Joshua Emele <jemele@chromium.org>