UPSTREAM: Bluetooth: mt7921s: Support wake on bluetooth
Enable wake on bluetooth on mt7921s that can be supported since the
firmware with version 20211129211059 was added, and the patch would
not cause any harm even when the old firmware is applied.

The patch was tested by setting up an HID or HOGP profile to connect a
Bluetooth keyboard and mouse, then putting the system to suspend, then
trying to wake up the system by moving the Bluetooth keyboard or mouse,
and then checking if the system can wake up and be brought back to
the normal state.

Co-developed-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Mark Chen <mark-yw.chen@mediatek.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit ce64b3e949195768361b0e4111e0837e5e46930d)

Conflicts:
	drivers/bluetooth/btmtk.h

BUG=none
TEST=build

Signed-off-by: Zhengping Jiang <jiangzp@google.com>