FROMLIST: x86/msr-index: Add MSRs for Key Locker internal wrapping key
The CPU state that contains the internal wrapping key is in the same power
domain as the cache. So any sleep state that would invalidate the cache
(like S3) also invalidates the state of the wrapping key.

But, since the state is inaccessible to software, it needs a special
mechanism to save and restore the key during deep sleep.

A set of new MSRs are provided as an abstract interface to save and restore
the wrapping key, and to check the key status.

The wrapping key is saved in a platform-scoped state of non-volatile media.
The backup itself and its path from the CPU are encrypted and integrity
protected.

But this storage's non-volatility is not architecturally guaranteed across
off states, such as S5 and G3.

The MSRs will be used to back up the key for S3/4 sleep states. Then the
kernel code writes one of them to restore the key in each CPU state.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Cc: x86@kernel.org
Cc: linux-kernel@vger.kernel.org
(am from https://patchwork.kernel.org/project/linux-crypto/patch/20220112211258.21115-6-chang.seok.bae@intel.com/)

BUG=b:198473496
TEST=compilation, tcrypt tests

Signed-off-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Change-Id: I085d0507f4a183f96150e46fb6e17cae7be31783
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3373767
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>