CHROMIUM: ath11k: wcn6855: update FW dir for hw version 2.1
WCN6855 hardware version 2.1 and 2.0 are supported in the same Firmware
but since the fw.dir is assigned as wcn6855/hw2.1, FW load fails on
hw2.1. As a solution, assign fw.dir for hw2.1 as wcn6855/hw2.0. An
upstream fix is requested for this kind of case in b/215198143

BUG=b:215198143, b:214319381
TEST=Verified on Nipperkin

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: Icdb0fcf38e9808568a83590613bc2738c52d23ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3399408
Reviewed-by: Joshua Emele <jemele@chromium.org>
Commit-Queue: Alan Chuang <alanchuang@chromium.org>