CHROMIUM: ASoC: SOF: Intel: pci-tgl: unblock S5 entry if DMA stop has failed
If system shutdown has not been completed cleanly, it is possible the
DMA stream shutdown has not been done, or was not clean.

If this is the case, Intel TGL/ADL HDA platforms may fail to shutdown
cleanly due to pending HDA DMA transactions. To avoid this, detect this
scenario in the shutdown callback, and perform an additional controller
reset. This has been tested to unblock S5 entry if this condition is
hit.

Co-developed-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>

This is patch to ensure reboot scripts doesnt fail often.
As discussed in bug, this is temproary fix till upstream fix is ready.
PR posted in https://github.com/thesofproject/linux/pull/3388

BUG=b:214665770
TEST=Verify firmwareconsecutiveBoot runs for 2500 cycles

Signed-off-by: Sathya Prakash M R <sathya.prakash.m.r@intel.com>
Change-Id: Ie872e1435fa47469ff02d86632f443bff58a4559
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3410448
Reviewed-by: Archana Patni <archana.patni@intel.corp-partner.google.com>
Tested-by: Archana Patni <archana.patni@intel.corp-partner.google.com>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>